Aprimorando o controle de login corrigindo alguns corner-cases.

+++ src/model/Specialist.java
@@ -10,11 +10,11 @@ import javax.persistence.Id;
 public class Specialist {
 
 	@Id
-	private int id;	
+	private long id;	
 	@Id
 	private String especialidade;
 	
-	public Specialist(int id,String especialidade){
+	public Specialist(long id, String especialidade){
 		this.id = id;
 		this.especialidade = especialidade;
 	}
